requirements.txt
rich
miservice_fork
requests
revChatGPT
openai
